backend/requirements.txt
fastapi~=0.104.1
SQLAlchemy~=2.0.23
uvicorn>=0.24
PyMySQL
python-dotenv
alembic
